requirements.txt
fastapi
uvicorn
python-dotenv
line-bot-sdk
psycopg2-binary
sqlalchemy